Beginning of the API
Not working fully yet, but it can send a query request.The problem right now seems to be the asynchronous access to "result" in JsonCcov while a query request is being granted preventing the result from being returned with a value other than undefined. The API works as follows:

1) Create a JsonCcov object.
2) Create a new Query object that represents the query that should be requested. 
3) Set the JsonCcov object to the query. (Multiple queries at once will also be supported).
4) Perform the query and retrieve the results in the form of a JSON.

Currently, there is only one Query for files touched by a given test.